+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/ArtemisTcpTransport.kt
@@ -1,44 +1,30 @@
-package net.corda.nodeapi
+package net.corda.nodeapi.internal
 
 import net.corda.core.messaging.ClientRpcSslOptions
 import net.corda.core.serialization.internal.nodeSerializationEnv
 import net.corda.core.utilities.NetworkHostAndPort
-import net.corda.nodeapi.internal.config.SSLConfiguration
-import net.corda.nodeapi.internal.requireOnDefaultFileSystem
+import net.corda.nodeapi.BrokerRpcSslOptions
+import net.corda.nodeapi.internal.config.CertificateStore
+import net.corda.nodeapi.internal.config.FileBasedCertificateStoreSupplier
+import net.corda.nodeapi.internal.config.SslConfiguration
+import net.corda.nodeapi.internal.config.MutualSslConfiguration
 import org.apache.activemq.artemis.api.core.TransportConfiguration
 import org.apache.activemq.artemis.core.remoting.impl.netty.NettyConnectorFactory
 import org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants
 import java.nio.file.Path
 
-/** Class to set Artemis TCP configuration options. */
+// This avoids internal types from leaking in the public API. The "external" ArtemisTcpTransport delegates to this internal one.
 class ArtemisTcpTransport {
     companion object {
-        /**
-         * Corda supported TLS schemes.
-         * <p><ul>
-         * <li>TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256
-         * <li>TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256
-         * <li>TLS_DHE_RSA_WITH_AES_128_GCM_SHA256
-         * </ul></p>
-         * As shown above, current version restricts enabled TLS cipher suites to:
-         * AES128 using Galois/Counter Mode (GCM) for the block cipher being used to encrypt the message stream.
-         * SHA256 as message authentication algorithm.
-         * Ephemeral Diffie Hellman key exchange for advanced forward secrecy. ECDHE is preferred, but DHE is also
-         * supported in case one wants to completely avoid the use of ECC for TLS.
-         * ECDSA and RSA for digital signatures. Our self-generated certificates all use ECDSA for handshakes,
-         * but we allow classical RSA certificates to work in case one uses external tools or cloud providers or HSMs
-         * that do not support ECC certificates.
-         */
         val CIPHER_SUITES = listOf(
                 "TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256",
                 "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256",
                 "TLS_DHE_RSA_WITH_AES_128_GCM_SHA256"
         )
 
-        /** Supported TLS versions, currently TLSv1.2 only. */
         val TLS_VERSIONS = listOf("TLSv1.2")
 
-        private fun defaultArtemisOptions(hostAndPort: NetworkHostAndPort) = mapOf(
+        internal fun defaultArtemisOptions(hostAndPort: NetworkHostAndPort) = mapOf(
                 // Basic TCP target details.
                 TransportConstants.HOST_PROP_NAME to hostAndPort.host,
                 TransportConstants.PORT_PROP_NAME to hostAndPort.port,
@@ -54,18 +40,47 @@ class ArtemisTcpTransport {
                 //hick-ups under high load (CORDA-1336)
                 TransportConstants.DIRECT_DELIVER to false)
 
-        private val defaultSSLOptions = mapOf(
+        internal val defaultSSLOptions = mapOf(
                 TransportConstants.ENABLED_CIPHER_SUITES_PROP_NAME to CIPHER_SUITES.joinToString(","),
                 TransportConstants.ENABLED_PROTOCOLS_PROP_NAME to TLS_VERSIONS.joinToString(","))
 
-        private fun SSLConfiguration.toTransportOptions() = mapOf(
+        private fun SslConfiguration.toTransportOptions(): Map<String, Any> {
+
+            val options = mutableMapOf<String, Any>()
+            (keyStore to trustStore).addToTransportOptions(options)
+            return options
+        }
+
+        private fun Pair<FileBasedCertificateStoreSupplier?, FileBasedCertificateStoreSupplier?>.addToTransportOptions(options: MutableMap<String, Any>) {
+
+            val keyStore = first
+            val trustStore = second
+            keyStore?.let {
+                with (it) {
+                    path.requireOnDefaultFileSystem()
+                    options.putAll(get().toKeyStoreTransportOptions(path))
+                }
+            }
+            trustStore?.let {
+                with (it) {
+                    path.requireOnDefaultFileSystem()
+                    options.putAll(get().toTrustStoreTransportOptions(path))
+                }
+            }
+        }
+
+        private fun CertificateStore.toKeyStoreTransportOptions(path: Path) = mapOf(
                 TransportConstants.SSL_ENABLED_PROP_NAME to true,
                 TransportConstants.KEYSTORE_PROVIDER_PROP_NAME to "JKS",
-                TransportConstants.KEYSTORE_PATH_PROP_NAME to sslKeystore,
-                TransportConstants.KEYSTORE_PASSWORD_PROP_NAME to keyStorePassword,
+                TransportConstants.KEYSTORE_PATH_PROP_NAME to path,
+                TransportConstants.KEYSTORE_PASSWORD_PROP_NAME to password,
+                TransportConstants.NEED_CLIENT_AUTH_PROP_NAME to true)
+
+        private fun CertificateStore.toTrustStoreTransportOptions(path: Path) = mapOf(
+                TransportConstants.SSL_ENABLED_PROP_NAME to true,
                 TransportConstants.TRUSTSTORE_PROVIDER_PROP_NAME to "JKS",
-                TransportConstants.TRUSTSTORE_PATH_PROP_NAME to trustStoreFile,
-                TransportConstants.TRUSTSTORE_PASSWORD_PROP_NAME to trustStorePassword,
+                TransportConstants.TRUSTSTORE_PATH_PROP_NAME to path,
+                TransportConstants.TRUSTSTORE_PASSWORD_PROP_NAME to password,
                 TransportConstants.NEED_CLIENT_AUTH_PROP_NAME to true)
 
         private fun ClientRpcSslOptions.toTransportOptions() = mapOf(
@@ -81,34 +96,39 @@ class ArtemisTcpTransport {
                 TransportConstants.KEYSTORE_PASSWORD_PROP_NAME to keyStorePassword,
                 TransportConstants.NEED_CLIENT_AUTH_PROP_NAME to false)
 
-        private val acceptorFactoryClassName = "org.apache.activemq.artemis.core.remoting.impl.netty.NettyAcceptorFactory"
-        private val connectorFactoryClassName = NettyConnectorFactory::class.java.name
+        internal val acceptorFactoryClassName = "org.apache.activemq.artemis.core.remoting.impl.netty.NettyAcceptorFactory"
+        internal val connectorFactoryClassName = NettyConnectorFactory::class.java.name
 
-        fun p2pAcceptorTcpTransport(hostAndPort: NetworkHostAndPort, config: SSLConfiguration?, enableSSL: Boolean = true): TransportConfiguration {
+        fun p2pAcceptorTcpTransport(hostAndPort: NetworkHostAndPort, config: MutualSslConfiguration?, enableSSL: Boolean = true): TransportConfiguration {
+
+            return p2pAcceptorTcpTransport(hostAndPort, config?.keyStore, config?.trustStore, enableSSL = enableSSL)
+        }
+
+        fun p2pConnectorTcpTransport(hostAndPort: NetworkHostAndPort, config: MutualSslConfiguration?, enableSSL: Boolean = true): TransportConfiguration {
+
+            return p2pConnectorTcpTransport(hostAndPort, config?.keyStore, config?.trustStore, enableSSL = enableSSL)
+        }
+
+        fun p2pAcceptorTcpTransport(hostAndPort: NetworkHostAndPort, keyStore: FileBasedCertificateStoreSupplier?, trustStore: FileBasedCertificateStoreSupplier?, enableSSL: Boolean = true): TransportConfiguration {
+
             val options = defaultArtemisOptions(hostAndPort).toMutableMap()
-
-            if (config != null && enableSSL) {
-                config.sslKeystore.requireOnDefaultFileSystem()
-                config.trustStoreFile.requireOnDefaultFileSystem()
+            if (enableSSL) {
                 options.putAll(defaultSSLOptions)
-                options.putAll(config.toTransportOptions())
+                (keyStore to trustStore).addToTransportOptions(options)
             }
             return TransportConfiguration(acceptorFactoryClassName, options)
         }
 
-        fun p2pConnectorTcpTransport(hostAndPort: NetworkHostAndPort, config: SSLConfiguration?, enableSSL: Boolean = true): TransportConfiguration {
-            val options = defaultArtemisOptions(hostAndPort).toMutableMap()
+        fun p2pConnectorTcpTransport(hostAndPort: NetworkHostAndPort, keyStore: FileBasedCertificateStoreSupplier?, trustStore: FileBasedCertificateStoreSupplier?, enableSSL: Boolean = true): TransportConfiguration {
 
-            if (config != null && enableSSL) {
-                config.sslKeystore.requireOnDefaultFileSystem()
-                config.trustStoreFile.requireOnDefaultFileSystem()
+            val options = defaultArtemisOptions(hostAndPort).toMutableMap()
+            if (enableSSL) {
                 options.putAll(defaultSSLOptions)
-                options.putAll(config.toTransportOptions())
+                (keyStore to trustStore).addToTransportOptions(options)
             }
             return TransportConfiguration(connectorFactoryClassName, options)
         }
 
-        /** [TransportConfiguration] for RPC TCP communication - server side. */
         fun rpcAcceptorTcpTransport(hostAndPort: NetworkHostAndPort, config: BrokerRpcSslOptions?, enableSSL: Boolean = true): TransportConfiguration {
             val options = defaultArtemisOptions(hostAndPort).toMutableMap()
 
@@ -120,8 +140,6 @@ class ArtemisTcpTransport {
             return TransportConfiguration(acceptorFactoryClassName, options)
         }
 
-        /** [TransportConfiguration] for RPC TCP communication
-         * This is the Transport that connects the client JVM to the broker. */
         fun rpcConnectorTcpTransport(hostAndPort: NetworkHostAndPort, config: ClientRpcSslOptions?, enableSSL: Boolean = true): TransportConfiguration {
             val options = defaultArtemisOptions(hostAndPort).toMutableMap()
 
@@ -133,19 +151,16 @@ class ArtemisTcpTransport {
             return TransportConfiguration(connectorFactoryClassName, options)
         }
 
-        /** Create as list of [TransportConfiguration]. **/
         fun rpcConnectorTcpTransportsFromList(hostAndPortList: List<NetworkHostAndPort>, config: ClientRpcSslOptions?, enableSSL: Boolean = true): List<TransportConfiguration> = hostAndPortList.map {
             rpcConnectorTcpTransport(it, config, enableSSL)
         }
 
-        fun rpcInternalClientTcpTransport(hostAndPort: NetworkHostAndPort, config: SSLConfiguration): TransportConfiguration {
+        fun rpcInternalClientTcpTransport(hostAndPort: NetworkHostAndPort, config: SslConfiguration): TransportConfiguration {
             return TransportConfiguration(connectorFactoryClassName, defaultArtemisOptions(hostAndPort) + defaultSSLOptions + config.toTransportOptions())
         }
 
-        fun rpcInternalAcceptorTcpTransport(hostAndPort: NetworkHostAndPort, config: SSLConfiguration): TransportConfiguration {
+        fun rpcInternalAcceptorTcpTransport(hostAndPort: NetworkHostAndPort, config: SslConfiguration): TransportConfiguration {
             return TransportConfiguration(acceptorFactoryClassName, defaultArtemisOptions(hostAndPort) + defaultSSLOptions + config.toTransportOptions())
         }
     }
 }
-
-data class BrokerRpcSslOptions(val keyStorePath: Path, val keyStorePassword: String)